New theme prop for FlowNavigation (#194)
* New theme prop for FLowNavigation

I've exposed a new prop for FlowNavigation to change the theme - I've added a common themes file as well. At the moment the only change in FlowNavigation depending on the theme is the logo

* Hides stepper on FlowNavigation dark mode

I've changed the themes to lowercase to match sizes, and I've also made the background go dark in the docs to better show the inversed logo.

* Changes from review

- Used classNames instead of conditionals in template strings
- Changed instances of undefined to null in docs
- Using oneOf in component
- Updated tests

* Added TODO, updated changelog, package.json, package-lock.json

* Bump version again

* Fixed theme plurality