feat: add a layout service for converting dynamic flow steps without layouts to DynamicLayout format
### Context
Layout is not a required property of dynamic flow steps so we need to handle layout of two different approaches. Rather than do this, we will simply map steps without layouts to layout format using a service. From then on, we can treat all dynamic flow steps as layouts and we need only one approach for the layout of our template.

### Changes
We add a service that maps three step types (form, decision & final) to a dynamic layout specification

### Considerations
The DynamicFlow component that will use the service is yet to be written.

We do not mimic the logic from V2 that narrows/widens the form based on the width of the fields inside the form, as in V3, it is not straightforward to interrogate the fields.

We do not convert "upload steps" into side by side layouts, as in V2. The intent is to add a layout to the step to achieve these more complex layouts.